refactor(analysis): import the analysis feature via its FSD barrel
Map.tsx and VectorAreasLayer.tsx now import AnalysisCTA / useSelectionStore from
@/src/features/analysis instead of reaching into the old feature internals.
