+++ app/components/Map.tsx
@@ -25,7 +25,7 @@ import DynamicTileLayers, {
 import VectorDataLayers from "./map/layers/VectorDataLayers";
 import AoiVectorTileLayers from "./map/layers/AoiVectorTileLayers";
 import SelectAreaLayer from "./map/layers/select-area-layer";
-import { AnalysisCTA } from "@/app/features/analysis/ui/AnalysisCTA";
+import { AnalysisCTA } from "@/src/features/analysis";
 import { useLegendHook } from "@/app/components/legend/useLegendHook";
 import GeoJsonLayers from "./map/layers/GeoJsonLayers";
 import PendingDrawArea from "./map/layers/PendingDrawArea";
